complete norm exept minirt.h
