+++ src/cylinder.c
@@ -6,7 +6,7 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/03 12:11:30 by susami            #+#    #+#             */
-/*   Updated: 2023/02/08 20:58:02 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/08 20:59:05 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
