+++ src/hittable_list.c
@@ -6,13 +6,14 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 12:23:52 by susami            #+#    #+#             */
-/*   Updated: 2023/02/08 19:37:58 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/08 21:32:51 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-bool	hittable_list_hit(const t_hittable *head, const t_ray *r, t_range tr, t_hit_record *rec)
+bool	hittable_list_hit(
+		const t_hittable *head, const t_ray *r, t_range tr, t_hit_record *rec)
 {
 	bool			hit_any;
 	FLOAT			closest_t;
