+++ src/ray_trace.c
@@ -6,7 +6,7 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 10:43:37 by susami            #+#    #+#             */
-/*   Updated: 2023/02/08 19:39:37 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/08 23:57:34 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,27 +31,25 @@ t_color	diffuse_light(t_hit_record *rec, t_lighting *lighting)
 t_color	specular_light(const t_ray *ray, t_hit_record *rec,
 		t_lighting *lighting)
 {
-	FLOAT		nldot;
-	FLOAT		vrdot;
-	FLOAT		vrdot_pow_alpha;
-	t_vec		reflec;
-	t_vec		b;
-	t_vec		view;
-	t_color		ki;
+	t_dot	dot;	
+	t_vec	reflec;
+	t_vec	b;
+	t_vec	view;
+	t_color	ki;
 
-	nldot = vec_dot(lighting->direction, rec->normal);
-	b = vec_scalar_mul(nldot * 2, rec->normal);
+	dot.nl = vec_dot(lighting->direction, rec->normal);
+	b = vec_scalar_mul(dot.nl * 2, rec->normal);
 	reflec = vec_sub(b, lighting->direction);
 	view = vec_scalar_mul(-1, ray->direction);
 	view = vec_unit(view);
-	vrdot = vec_dot(view, reflec);
-	if (nldot <= 0 || vrdot <= 0)
+	dot.vr = vec_dot(view, reflec);
+	if (dot.nl <= 0 || dot.vr <= 0)
 		return (color_new(0, 0, 0));
-	vrdot_pow_alpha = pow(vrdot, rec->hittable_ptr->shineness);
+	dot.vr_pow_alpha = pow(dot.vr, rec->hittable_ptr->shineness);
 	ki = vec_mul(rec->hittable_ptr->specular_factor, lighting->intensity);
 	return (vec_mul(
 			ki,
-			color_new(vrdot_pow_alpha, vrdot_pow_alpha, vrdot_pow_alpha)));
+			color_new(dot.vr_pow_alpha, dot.vr_pow_alpha, dot.vr_pow_alpha)));
 }
 
 t_lighting	lighting_at(const t_light_source *self, t_vec pos)
@@ -72,7 +70,8 @@ bool	is_shadow(const t_hit_record *rec, const t_lighting *l,
 	t_hit_record	temp_rec;
 
 	shadow_ray = (t_ray){.direction = l->direction, .origin = rec->p};
-	return (hit(&scene->list, &shadow_ray, range_new(C_EPSILON, l->distance), &temp_rec));
+	return (hit(&scene->list, &shadow_ray,
+			range_new(C_EPSILON, l->distance), &temp_rec));
 }
 
 t_color	ray_trace(const t_ray *r, t_scene *scene)
