+++ src/translate1.c
@@ -1,16 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   translate.c                                        :+:      :+:    :+:   */
+/*   translate1.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 11:37:14 by katakagi          #+#    #+#             */
-/*   Updated: 2023/02/07 11:08:21 by susami           ###   ########.fr       */
+/*   Updated: 2023/02/08 21:02:40 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
+
+void	translate_ambient(t_scene *scene, t_element *elem);
+void	translate_light(t_scene *scene, t_element *elem);
+void	translate_camera(t_scene *scene, t_element *elem);
 
 void	set_material_info(t_hittable *obj, t_color color)
 {
@@ -18,12 +22,6 @@ void	set_material_info(t_hittable *obj, t_color color)
 	obj->diffuse_factor = vec_scalar_mul(0.69 / 255, color);
 	obj->specular_factor = vec_scalar_mul(0.3 / 255, color);
 	obj->shineness = 8.0;
-}
-
-void	translate_ambient(t_scene *scene, t_element *elem)
-{
-	scene->ambient_intensity = vec_scalar_mul(
-			elem->ambient_lightning_ratio / 255, elem->color);
 }
 
 void	translate_sphere(t_scene *scene, t_element *elem)
@@ -39,21 +37,6 @@ void	translate_sphere(t_scene *scene, t_element *elem)
 	cur->center = elem->sp_center;
 	cur->radius = elem->sp_diameter;
 	set_material_info(cur, elem->color);
-}
-
-void	translate_light(t_scene *scene, t_element *elem)
-{
-	scene->light_source.intensity = vec_scalar_mul(
-			elem->light_brightness_ratio / 255,
-			elem->color);
-	scene->light_source.position = elem->light_point;
-}
-
-void	translate_camera(t_scene *scene, t_element *elem)
-{
-	scene->camera.eye_position = elem->view_point;
-	scene->camera.look_at_direction = elem->orientation;
-	scene->camera.hfov = elem->hfov;
 }
 
 void	translate_plane(t_scene *scene, t_element *elem)
