+++ src/vec.h
@@ -6,7 +6,7 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/31 14:15:34 by susami            #+#    #+#             */
-/*   Updated: 2023/02/08 20:39:21 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/09 01:21:28 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,12 @@
 # define VEC_H
 
 # define FLOAT double
-# include "for_norm.h"
+// # include "for_norm.h"
 
 typedef struct s_vec	t_vec;
 typedef struct s_vec	t_point;
 typedef struct s_vec	t_color;
+typedef struct s_range	t_range;
 
 t_vec	vec_mul(t_vec a, t_vec b);
 t_vec	vec_scalar_mul(FLOAT t, t_vec v);
@@ -28,6 +29,11 @@ struct s_vec {
 	FLOAT	x;
 	FLOAT	y;
 	FLOAT	z;
+};
+
+struct s_range {
+	FLOAT	min;
+	FLOAT	max;
 };
 
 t_vec	vec_new(FLOAT x, FLOAT y, FLOAT z);
